test(checker): split JS constructor property tests
